pyproject.toml
[tool.poetry]
name = "azure-pages-router"
version = "0.1.0"
description = "Azure Pages Router"
authors = ["Cleison Amorim <cca5@cin.ufpe.br>"]

[tool.poetry.dependencies]
python = "^3.9"
# commons-python = { git = "https://github.com/amorim-cleison/commons-python.git", branch = "develop" }
# commons-python = { path = "../commons-python/", develop = false }
azure-functions = "^1.13.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
yapf = "^0.30.0"
pytest = "^6.1.1"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
